% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{.ConvertCharToRaw}
\alias{.ConvertCharToRaw}
\title{Convert character string to UUID raw representation}
\usage{
.ConvertCharToRaw(x)
}
\arguments{
\item{x}{A string; may be in upper or lower-case}
}
\value{
Returns a raw vector of 16 bytes
}
\description{
Convert character string to UUID raw representation
}
\examples{
.ConvertCharToRaw("00000000-0000-0000-0000-000000000000")
}
