% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{.ConvertRawToChar}
\alias{.ConvertRawToChar}
\title{Convert raw 16-byte vector to UUID character representation}
\usage{
.ConvertRawToChar(x)
}
\arguments{
\item{x}{A raw vector, must be 16 byte}
}
\value{
Returns a well-formed string representation of the UUID.
}
\description{
Convert raw 16-byte vector to UUID character representation
}
\examples{
.ConvertRawToChar(raw(16))
}
