% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\docType{class}
\name{uuid-class}
\alias{uuid-class}
\title{UUID Class}
\description{
UUID Class
}
\section{Slots}{

\describe{
\item{\code{char_repr}}{Character representation of the UUID}

\item{\code{raw}}{Raw (bytes) representation of the UUID}
}}

