% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate UUID strings}
\usage{
uuid(n = 1, version = c("1", "3", "4", "5"))
}
\arguments{
\item{n}{Output length}

\item{version}{UUID version}
}
\value{
A character vector
}
\description{
Generate a character vector of UUID strings of given length
}
